plotter-gcode-webapp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python==4.10.0.84
Pillow==10.4.0
svgpathtools==1.6.1
numpy==2.1.1
